requirements.txt
alembic==1.11.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.119.1
filelock==3.19.1
fonttools==4.60.1
fredapi==0.5.2
frozendict==2.4.6
fsspec==2025.9.0
h11==0.14.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
idna==3.11
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multitasking==0.0.12
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.23.1
protobuf==6.33.0
pycparser==2.23
pydantic==1.10.24
pydantic_core==2.41.4
pyparsing==3.2.5
pytest==7.4.0
pytest-asyncio==0.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
starlette==0.48.0
sympy==1.14.0
ta==0.11.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yfinance==0.2.66
zipp==3.23.0
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26
yfinance>=0.2.40
fredapi>=0.5
torch>=2.3
scikit-learn>=1.5
ta>=0.11
matplotlib>=3.8
joblib>=1.4
PyYAML>=6.0.1
